- Modificacao das classes de entidades. - Movido para o pacote modelo as entidades nao persistentes - Corrigida as anotacoes das classes persistentes.

+++ Desenvolvimento/avadoc-022013/avadoc-server/src/main/java/br/ufg/inf/avadoc/entity/IAtividade.java
@@ -14,6 +14,6 @@ public interface IAtividade {
 	 * 
 	 * @return pontos
 	 */
-	public int getPontos();
+	public double getPontos();
 
 }
